[Refactor] 오타수정, 파일 분리
