+++ MateRunner/MateRunner/Application/SceneDelegate.swift
@@ -7,18 +7,6 @@
 
 import UIKit
 
-enum NotificationCenterKey {
-    static let invitationDidRecieve = Notification.Name("invitationDidRecieve")
-    static let invitation = "invitation"
-    static let sessionID = "sessionId"
-    static let host = "host"
-    static let mate = "mate"
-    static let inviteTime = "inviteTime"
-    static let mode = "mode"
-    static let targetDistance = "targetDistance"
-    static let sender = "sender"
-}
-
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     
     var window: UIWindow?
