+++ MateRunner/MateRunner/Presentation/Common/Coordinator/Protocol/InvitationReceivable.swift
@@ -1,0 +1,14 @@
+//
+//  InvitationReceivable.swift
+//  MateRunner
+//
+//  Created by 전여훈 on 2021/11/28.
+//
+
+import Foundation
+
+protocol InvitationReceivable: AnyObject {
+    func invitationDidReceive(_ notification: Notification)
+    func invitationDidAccept(with settingData: RunningSetting)
+    func invitationDidReject()
+}
