+++ MateRunner/MateRunner/Presentation/Common/Coordinator/Protocol/TabBarCoordinator.swift
@@ -7,7 +7,7 @@
 
 import UIKit
 
-protocol TabBarCoordinator: Coordinator, InvitationRecievable {
+protocol TabBarCoordinator: Coordinator, InvitationReceivable {
     var tabBarController: UITabBarController { get set }
     func selectPage(_ page: TabBarPage)
     func setSelectedIndex(_ index: Int)
